NearBeach/requirements.txt
# requirement.txt
Django>=4.2.15
django-rest-framework
django-two-factor-auth
django-two-factor-auth[phonenumberslite]
simplejson
pillow>=10.0.1
urllib3>2.2.2
boto3==1.35.99
azure-storage-blob
azure-identity>=1.16.1

pip>=19.2 # not directly required, pinned by Snyk to avoid a vulnerability
requests>=2.32.2 # not directly required, pinned by Snyk to avoid a vulnerability
